[MODEL-24523] -- refactor
