+++ src/datarobot_genai/drmcputils/exceptions.py
@@ -103,3 +103,7 @@ class ToolError(Exception):
     def __repr__(self) -> str:
         """Return a string representation of the exception."""
         return f"{self.__class__.__name__}({self.message!r}, kind={self.kind!r})"
+
+
+class AudienceClaimValidationError(Exception):
+    pass
